requirements.txt
google-genai
python-dotenv
uvicorn
requests
fastapi
pydantic